Merge dev into master for deploying to Heroku.

+++ server/routes/order.router.js
@@ -6,15 +6,18 @@ const {
 const pool = require('../modules/pool');
 
 router.get('/', rejectUnauthenticated, async (req, res) => {
+  const accessLevel = req.user.access_level;
+  if (accessLevel < 10) {
+    res.sendStatus(401);
+    return;
+  }
   const conn = await pool.connect();
   try {
-    const query = {};
-    query.text = `SELECT "order".*, account."name", account.email,
-      profile.household_id, profile.last_pickup FROM "order"
+    const result = await conn.query(`SELECT "order".*, account."name", account.email,
+      profile.household_id, profile.latest_order FROM "order"
       LEFT JOIN account ON "order".account_id = account.id
-      LEFT JOIN profile ON account.id = profile.account_id;`;
-    query.values = [];
-    const result = await conn.query(query.text, query.values);
+      LEFT JOIN profile ON account.id = profile.account_id;`);
+    conn.release();
     res.status(200).send(result.rows);
   } catch (error) {
     console.log('Error GET /api/order', error);
@@ -23,16 +26,21 @@ router.get('/', rejectUnauthenticated, async (req, res) => {
 });
 
 router.get('/active', rejectUnauthenticated, async (req, res) => {
+  const accessLevel = req.user.access_level;
+  if (accessLevel < 10) {
+    res.sendStatus(401);
+    return;
+  }
   const conn = await pool.connect();
   try {
-    const query = {};
-    query.text = `SELECT "order".*, account."name", profile.household_id, profile.last_pickup FROM "order"
-      LEFT JOIN account ON "order".account_id = account.id
-      LEFT JOIN profile ON account.id = profile.account_id 
-      WHERE checkout_at IS NULL 
-      ORDER BY checkin_at DESC;`;
-    query.values = [];
-    const result = await conn.query(query.text, query.values);
+    const result = await conn.query(`SELECT "order".*, account."name",
+                                    profile.household_id, profile.latest_order FROM "order"
+                                    LEFT JOIN account ON "order".account_id = account.id
+                                    LEFT JOIN profile ON account.id = profile.account_id 
+                                    WHERE cast(checkin_at as date) = CURRENT_DATE
+                                    AND checkout_at IS NULL
+                                    ORDER BY checkin_at DESC;`);
+    conn.release();
     res.status(200).send(result.rows);
   } catch (error) {
     console.log('Error GET /api/order/active', error);
@@ -41,18 +49,40 @@ router.get('/active', rejectUnauthenticated, async (req, res) => {
 });
 
 router.get('/complete/today', rejectUnauthenticated, async (req, res) => {
+  const accessLevel = req.user.access_level;
+  if (accessLevel < 10) {
+    res.sendStatus(401);
+    return;
+  }
+  const conn = await pool.connect();
+  try {
+    const result = await conn.query(`SELECT "order".*, account."name",
+                                    profile.household_id, profile.latest_order FROM "order"
+                                    LEFT JOIN account ON "order".account_id = account.id
+                                    LEFT JOIN profile ON account.id = profile.account_id 
+                                    WHERE cast(checkin_at as date) = CURRENT_DATE
+                                    AND checkout_at IS NOT NULL
+                                    ORDER BY checkin_at DESC;`);
+    conn.release();
+    res.status(200).send(result.rows);
+  } catch (error) {
+    console.log('Error GET /api/order/active', error);
+    res.sendStatus(500);
+  }
+});
+
+router.get('/client-order-status', async (req, res) => {
+  const id = req.user.id;
   const conn = await pool.connect();
   try {
     const query = {};
-    query.text = `SELECT "order".*, account."name", profile.household_id, profile.last_pickup FROM "order"
-      LEFT JOIN account ON "order".account_id = account.id
-      LEFT JOIN profile ON account.id = profile.account_id 
-        WHERE checkout_at 
-          BETWEEN NOW() - INTERVAL '24 HOURS' 
-          AND NOW()
-        ORDER BY checkout_at DESC;`;
-    query.values = [];
+    query.text = `SELECT "order".* FROM "order"
+                  WHERE cast(checkin_at as date) = CURRENT_DATE
+                  AND "order".account_id = $1
+                  ORDER BY checkin_at DESC;`;
+    query.values = [id];
     const result = await conn.query(query.text, query.values);
+    conn.release();
     res.status(200).send(result.rows);
   } catch (error) {
     console.log('Error GET /api/order/active', error);
@@ -62,18 +92,36 @@ router.get('/complete/today', rejectUnauthenticated, async (req, res) => {
 
 router.post('/', rejectUnauthenticated, async (req, res) => {
   const accountID = req.user.id;
+  const accessLevel = req.user.access_level;
 
   const locationID = req.body.location_id;
   const dietaryRestrictions = req.body.dietary_restrictions;
   const walkingHome = req.body.walking_home;
   const pregnant = req.body.pregnant;
   const childBirthday = req.body.child_birthday;
+  const snap = req.body.snap;
+  const pickupName = req.body.pickup_name;
+  const other = req.body.other;
+  let waitTimeMinutes = null;
 
+  if (accessLevel >= 10) {
+    try {
+      waitTimeMinutes = Number(req.body.wait_time_minutes);
+    } catch (error) {
+      res.sendStatus(400);
+      return;
+    }
+  }
+
+  // TODO only allow volunteers to specify the wait time.
   if (!locationID ||
-    !dietaryRestrictions ||
+    typeof dietaryRestrictions !== 'string' ||
     typeof walkingHome !== 'boolean' ||
-    typeof pregnant !== 'boolean'  ||
-    typeof childBirthday !== 'boolean'
+    typeof pregnant !== 'boolean' ||
+    typeof childBirthday !== 'boolean' ||
+    typeof snap !== 'boolean' ||
+    typeof pickupName !== 'string' ||
+    typeof other !== 'string'
   ) {
     res.sendStatus(400);
     return;
@@ -88,9 +136,13 @@ router.post('/', rejectUnauthenticated, async (req, res) => {
       dietary_restrictions,
       walking_home,
       pregnant,
-      child_birthday
+      child_birthday,
+      snap,
+      pickup_name,
+      other,
+      wait_time_minutes
       )
-      VALUES ($1, $2, $3, $4, $5, $6)
+      VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)
       RETURNING *;`;
     query.values = [
       accountID,
@@ -98,11 +150,16 @@ router.post('/', rejectUnauthenticated, async (req, res) => {
       dietaryRestrictions,
       walkingHome,
       pregnant,
-      childBirthday
+      childBirthday,
+      snap,
+      pickupName,
+      other,
+      waitTimeMinutes
     ];
     await conn.query('BEGIN');
     const result = await conn.query(query.text, query.values);
     await conn.query('COMMIT');
+    conn.release();
     res.status(201).send(result.rows[0]);
   } catch (error) {
     await conn.query('ROLLBACK');
@@ -112,17 +169,38 @@ router.post('/', rejectUnauthenticated, async (req, res) => {
 });
 
 router.put('/checkout/:id', async (req, res) => {
+  const accessLevel = req.user.access_level;
+  if (accessLevel < 10) {
+    res.sendStatus(401);
+    return;
+  }
+  let waitTimeMinutes;
+  try {
+    waitTimeMinutes = Number(req.body.wait_time_minutes);
+  } catch (error) {
+    res.sendStatus(400);
+    return;
+  }
+
   const conn = await pool.connect();
   try {
-    const query = {};
-    query.text = `UPDATE "order"
-      SET checkout_at = NOW()
+    const placeOrderQuery = {};
+    placeOrderQuery.text = `UPDATE "order"
+      SET checkout_at = NOW(), wait_time_minutes = $2
       WHERE id = $1
       RETURNING *;`;
-    query.values = [req.params.id];
+    placeOrderQuery.values = [req.params.id, waitTimeMinutes];
     await conn.query('BEGIN');
-    const result = await conn.query(query.text, query.values);
+    const result = await conn.query(placeOrderQuery.text, placeOrderQuery.values);
+    // do a second query that updates the user's profile to include the latest order
+    const updateProfileLatestOrderQuery = {};
+    updateProfileLatestOrderQuery.text = `UPDATE "profile"
+      SET latest_order = $1
+      WHERE account_id = $2;`;
+    updateProfileLatestOrderQuery.values = [result.rows[0].id, result.rows[0].account_id];
+    await conn.query(updateProfileLatestOrderQuery.text, updateProfileLatestOrderQuery.values);
     await conn.query('COMMIT');
+    conn.release();
     res.status(200).send(result.rows);
   } catch (error) {
     conn.query('ROLLBACK');
@@ -132,12 +210,18 @@ router.put('/checkout/:id', async (req, res) => {
 });
 
 router.delete('/:id', rejectUnauthenticated, async (req, res) => {
+  const accessLevel = req.user.access_level;
+  if (accessLevel < 100) {
+    res.sendStatus(401);
+    return;
+  }
   const conn = await pool.connect();
   try {
     const query = {};
     query.text = 'DELETE FROM "order" WHERE id = $1;';
     query.values = [req.params.id];
     await conn.query(query.text, query.values);
+    conn.release();
     res.sendStatus(204);
   } catch (error) {
     console.log('Error PUT /api/order/checkout/id', error);
